pyproject.toml
[project]
name = "omniccg"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.12.9"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "gitpython (>=3.1.45,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
